+++ include/StagingBufferPrefetcher.h
@@ -9,7 +9,7 @@ public:
     StagingBufferPrefetcher(char* staging_buffer,
                             unsigned long long int buffer_size,
                             int node_id,
-                            std::deque<int>* file_ends,
+                            std::deque<unsigned long long int> *file_ends,
                             std::mutex* staging_buffer_mutex,
                             std::condition_variable* staging_buffer_cond_var,
                             Sampler* sampler,
@@ -21,10 +21,10 @@ private:
     unsigned long long int buffer_size;
     int prefetch_batch = 0;
     int prefetch_offset = 0;
-    unsigned long staging_buffer_pointer = 0;
+    unsigned long long int staging_buffer_pointer = 0;
     int read_offset = 0;
     char* staging_buffer;
-    std::deque<int>* file_ends;
+    std::deque<unsigned long long int>* file_ends;
     std::mutex* staging_buffer_mutex;
     std::condition_variable* staging_buffer_cond_var;
     Sampler* sampler;
